ABI Monitoring: add bootstrap
Add the `bootstrap` script that is installing all the prerequisites that
are necessary to run the ABI utilities. In particular, the script takes
care of installing GitPython and all required dependencies for building
libabigail. It finally builds the abigail tools from source and
provides instructions to make them available in PATH.

Bug: 111449103
Co-authored-by: Jayant Chowdhary <jchowdhary@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Id44efe6874f7b51bc64abca9ac3d879612180f85