kernelscripts: Switch to recursive copy instead of cp --parents
Recent builds have started using the prebuilt cp binary that
is shipped with Android builds, as opposed to the host cp.
However, the cp binary does not properly implement the
--parents option correctly, and as such, files that
are copied with this option will not get copied correctly.

Since all of the files copied with the --parents options are all
under one directory, then use a recursive copy to copy the
files.

Change-Id: Icf9046067f195e4fc005372b4499bd513d116231
Signed-off-by: Isaac J. Manjarres <isaacm@codeaurora.org>