build.sh: add option to define abi definition location
If ABI_DEFINITION is set (usually within a build.config file), make
build.sh copy it over to the distribution. Since it is copied from the
sources it might not match the actual ABI of the binaries that are part
of the distribution. Finding this difference is subject to tests /
analysis run against the distribution for now. And making this file part
of the distribution actually enables this kind of analysis.

In case we are running build.sh as part of build_abi.sh, disable this
functionality as we are supposed to generate the abi definition in this
case.

Bug: 133501930
Change-Id: Idd49161a311b0e8995108a875f0c46d61965a2d7
Signed-off-by: Matthias Maennich <maennich@google.com>