abi: replace libabigail's --short-locs by postprocessing of the abi.xml
Shortening the file paths with --short-locs to just file names makes the
abi.xml output ambigous and the later abidiff analysis sensitive for
external influences that (still correct) change the result of the
binary. Hence, stop using --short-locs, but rather textual replace all
occurrences of the kernel source dir and - in a second step - everything
else that sneaked in (e.g. from the prebuilts or other repositories of
the checkout).

Bug: 135011693
Change-Id: I808a96e3e38365e54f5eb467587b8f5dfbc12825
Signed-off-by: Matthias Maennich <maennich@google.com>