build.sh: Set make flags and propagate make arguments
MAKEFLAGS are now enriched at the beginning of the script with a default
number of jobs (-j $(nproc)). This avoids specifying that option at
every make invocation.
An advantage is that also mrproper is be parallelized, and although its
operations are mostly io-bound, the total execution time can be strongly
decreased.

In addition to that, make arguments are propagated to all the make
commands.

Test: manual build
Bug: n/a
Change-Id: I717ac5ac3a92e377cee2766416825350cc4063f1
Signed-off-by: Alessio Balsini <balsini@android.com>