kernel/build: Append RTIC dtb in dtb image for boot header v2
With Boot image header version 2, dtb is no longer appended to
kernel image. dtb.img is generated separately. So concatenate RTIC
dtb to the dtb.img for the bootloader to read the rtic dtb and share
the measurement parameter to hypervisor. Make sure that the RTIC DTB 
rule added regardless of the presence of rtic_mp.dts. Instead, check
this during the rule execution (never fail if RTIC DTS is missing).

Change-Id: If2f093860b8a05aa2843173eaa49f85b86856b52