kernelscripts: Add snapshots of buildkernel.sh and kernel_definitions.mk
By default, the build.sh provides a mechanism to compile the kernel
as a standalone component, and is not helpful to compile from within
the Android tree. As a result, add the snapshots of kernel_definitions.mk
and buildkernel.sh files to support kernel compilation from within
the Android tree. The snaphot is taken from msm-4.19 branch as of
commit <0780fd437f12d401c> (kernel/build: remove wildcard entry for
updating makefile variable).

Change-Id: I63b3859a9c7d7b5bcf19320fe62235f432f9817a