kernel_definitions: Use TARGET_KERNEL_CROSS_COMPILE_PREFIX
TARGET_TOOLS_PREFIX is a new variable introduced in build environment
which keeps the location of the toolchain from the android source,
which was working for 64bit but giving unwind warnings for 32 bit
targets. To solve this let's make the toolchain path prefix change
in the target. BoardConfig file as per architecture(32/64) basis and
avoid making change in kernel.

Change-Id: If2efada56d81dfea921e93ded05d8ae30379d4c7