build.sh: Replace strip with objcopy
With the introduction of RELR, our legacy toolchains will fail to strip
binaries using these relocation types with an error like this:

vmlinux: unknown type [0x13] section `.relr.dyn'

We could respin binutils, but as we don't use those tools any more to
build the kernel, just use llvm-objcopy instead.

If this is an old build where OBJCOPY is unset, it assumes it can use
the binutils strip, we don't bother using binutils objcopy in that
situation.

Change-Id: Iffa6474e2058c5cad08b5576d3f8a80ec2a63064
Signed-off-by: Alistair Delva <adelva@google.com>