build.sh: Treat missing symbols from depmod as an error
Currently depmod isn't treating missing symbols as an error. This causes
the modules with missing symbols to fail to load. A missing symbols
warning will occur when an out-of-tree driver is compiled and missing
a symbol either from the kernel or some other kernel module.

This patch parses the output of depmod to detect any missing symbols. If
any missing symbols are found, then build.sh will return an error.

Signed-off-by: Will McVicker <willmcvicker@google.com>
Test: Remove symbol that an out-of-tree driver depends on
Change-Id: Ia74c11ea843cba6057c7be85889d6808446e5881