abi: Add script to extract an ABI whitelist
Add abi/extract_symbols as a means to extract the symbols in use from
kernel modules out of a dist (preferred) or build directory.
The script locations Kernel binaries (vmlinx, kernel modules) and
collects the exported and required symbols from them. It creates a list
of symbols that are export from vmlinux and (if
--include-module-exports is specified) and required (undefined) by any
kernel module. That list represents the symbols to potentially keep ABI
stable. The script can update or create such a whitelist (--whitelist).

A sample execution is

  $ build/abi/extract_symbols                        \
        out/android-mainline/dist/                   \  # the search dir
        --whitelist common/abi_gki_aarch64_whitelist

The script will also report symbols that are required by a module, but
are not defined in any kernel binary. That is a consistency check.

Bug: 145600453
Bug: 145692847
Change-Id: I0840125c5fc317d69f27aa53da16608176d935ad
Signed-off-by: Matthias Maennich <maennich@google.com>