build_abi.sh: omit impacted interfaces for the printed report
Using build_abi.sh with --print-report is a means to print a useful
message for commits. The list of impacted interfaces can become quite
lengthy and can be omitted in most cases as the impact can be seen from
the sheer number alone.

Hence, adjust diff_abi to also generate a short report and use this
report instead for printing the report.

Change-Id: Ib400c0d41a511dbf8be7dd9bbf94ff2bd23ba7a3
Signed-off-by: Matthias Maennich <maennich@google.com>