build_abi: build in separate (wiped out) OUT_DIR
Introduce the OUT_DIR_SUFFIX env variable to adjust the default OUT_DIR
set by _setup_env.sh. Further, make use of that when defining the
OUT_DIR for the ABI collection build.

When building for ABI analysis, we now build in a separate directory
(unless OUT_DIR is explicitly specified) and wipe that directory before
to ensure a clean / hermetic build and analysis.

Bug: 145543807
Change-Id: I11d70ffcf0634242b5711b293e0ca3475fed9bbf
Signed-off-by: Matthias Maennich <maennich@google.com>