kmi_defines: Significantly reduce CPU usage due to MP
Prior to these changes the CPU use time for a 4.19 allmodconfig
kernel was:
    real    1m53.605s
    user    4m29.334s
    sys     16m59.105s

Use of the chunk_size argumet to pool.map() causes very significant
system CPU usage time reduction to more than 13.5 minutes, but it
increases the real time by 35 seconds:
    real    2m37.096s
    user    4m32.222s
    sys     3m16.068s

To address the increase in real time, the kernel_component_factory()
invocation on "vmlinux.o" is run in its own dedicated process, because
the number of .o files involved for it very large compared to the
number of .o files for kernel modules.  The kernel_component_factory()
invocations for the kernel modules occur with a multiprocessing.Pool().

This CPU use times after separating the work for vmlinux.o is:
    real    1m32.939s
    user    4m28.241s
    sys     3m36.381s

Note the 20 second real time speedup from the code without chunk_size
specification, and the 1 minute speedup from the code with chunk_size
specification.  This last change does increase the system CPU time by
20 seconds, but it is uses over 13 minutes less of system time.

The numbers above are representative across multiple-runs, the speedups
are enough to merit the use of chunk_size and a dedicated process to
work on "vmlinux.o" as soon as possible and without that work
postponing the parallel work on the kernel modules.

Bug: 139885116
Change-Id: I14bc3941b457aac933e5f94fa638f8f49c4ffd6b
Signed-off-by: Matthias Maennich <maennich@google.com>