abi: extract_symbols: group symbols by requiring module
In order to make the symbol whitelists easier to maintain, group the
symbols by requiring module. A section with commonly required symbols
remains. The new output would look like:

  [abi_whitelist]
  # commonly used symbols
    func1
    var1

  # required by mymod.ko
    func2
    var2

Also drop the functionality to update whitelists. Now that GKI is
without modules, we do not need to pile on a central whitelists. The
addititive nature of contributed whitelists needs to be checked at
review time anyway. Also, updating commented whitelists is now becoming
infeasible for this tool.

Change-Id: I8dca8db15117a519004f69be55b9117a14807bd9
Signed-off-by: Matthias Maennich <maennich@google.com>