Add utilities to work on a hermetic Android Kernel Build
With working docker environment, do

  $ ./run.sh

That will build the container image and spawn a ACK build in the
container.

As a first measure, PATH is restricted to the prebuilts and some
selected host binaries. This does not stop the build from directly
accessing absolute paths and to use libraries, headers etc. But it is a
starting point.

Bug: 135922132
Change-Id: I6295f6b97d3c1c6c72fe840cd22a76b63d207727
Signed-off-by: Matthias Maennich <maennich@google.com>