build.sh: Be compatible with toybox gzip implementation
Commit eb6219c ("Expand support for creating a boot.img") introduced a
gzip command line that takes advantage of the -t switch to determine if
a given file is gzip compressed. However, the command line switch -t
appears to not be supported by the version of gzip that toybox provides
which is what the build system is using.

Instead of this extra step of determining whether the file we are
looking at is a gzip file, let's just try to extract it right away. If
that fails, we assume it's not a gzip file.

Fixes: commit eb6219c ("Expand support for creating a boot.img")
Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I6ce0f39316f29f5574f9efc49707688c7777edf8