build.sh: Set BOOT_IMAGE_HEADER_VERSION default to 3
Previously, build.sh assumed two different default values for
BOOT_IMAGE_HEADER_VERSION. With this change, we set it to 3 very early
to get consistent behaviour across the entire shell script. This also
avoids the following error message that we saw previously if
BOOT_IMAGE_HEADER_VERSION was not set.

./build/build.sh: line 489: [: : integer expression expected

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: Ic7aa7f854f08a132765f9481f6378fcf460c8429