build.sh: write abi.prop to contain ABI related properties
Depending on whether the build configuration defines a KMI definition
and/or whitelists, emit those into a .prop file during the build.

The file could contain more information (like KMI version) at a later
point. It can be picked up by testing infrastructure or by humans to
inspect the build result.

As part of this change, consolidate the variable definitions for
abi.xml and abi_whitelist.

Change-Id: I52723565cbe5d31d5b9c19348b34c6be0811a4e2
Signed-off-by: Matthias Maennich <maennich@google.com>