build.sh: Add option KMI_ENFORCED and propagate to abi.prop
Teach build.sh the option KMI_ENFORCED. It sole purpose as of now is to
propagate this option to the abi.prop file in DIST_DIR to be picked up
by infrastructure.

In theory, this could influence build_abi.sh as well, but the default of
build_abi.sh is to warn/fail in case of differences anyway right now.

Change-Id: Ibd98f975e934c01bfc358e6d5eecdf4d4d7f0908
Signed-off-by: Matthias Maennich <maennich@google.com>