build.sh: Simplify VENDOR_RAMDISK_BINARY handling
The code to concatenate ramdisks provides two features:

- Locate ramdisk image in either of two directories
- Identify type of ramdisk i.e. compressed or plain CPIO archive

Previously, this has been done in a for loop to allow for multiple
ramdisks. It turned out, though, that there is only a single ramdisk
(i.e.  VENDOR_RAMDISK_BINARY) for which such operations are necessary.
We can therefore get rid of the two for loops.

In addition, we are dropping support for VENDOR_RAMDISK_BINARY being a
path relative to ${DIST_DIR} as ${DIST_DIR} is an output directory and
appears to be an unlikely choice to store an input file.

Signed-off-by: Daniel Mentz <danielmentz@google.com>
Change-Id: I2e0a4193e02f031f3adec7a360e6f8a61ac063c1