extract_symbols: Add --gki-modules
Add a new option to extract_symbols to pass a list of GKI modules. When
set, all modules in that list will be removed from the list of standard
modules, and all symbols they export will be appended to the vmlinux
symbol list.

In other word, all GKI modules will be virtually merged into vmlinux
hence ensuring the vmlinux / GKI modules ABI is not captured in the new
symbol list.

Bug: 158655966
Signed-off-by: Quentin Perret <qperret@google.com>
Change-Id: I73704be1a94e3c42521e605c886f0b54c10f8408