blocklist: Replace all '-' to '_' in blocklist file
Modprobe doesn't differenciate between '-' and '_'. That is, the
output of 'modprobe -l' returns module names with '_' even though
the .ko file has '-' in the name. Since we compare this output
with the file name to check if the file is actually blocklisted in
vendor_modprobe script, we fail. Hence, convert all the module names in
the blocklist file to '_'. This change won't affect the modprobe -b.

Change-Id: Ia5e6910b533d4aa491a96b3f00bf1017780502c5
Signed-off-by: Prakruthi Deepak Heragu <pheragu@codeaurora.org>