kernel_definitions: Build the QGKI and GKI kernels in parallel
For $(PLATFORM)-qgki_defconfig, currently we build GKI and QGKI
kernels in a serial manner. To improve the build time, build the
kernels in parallel. For this, remove the 'make' target dependencies
between the two kernel builds and make sure that the defconfigs
for both the kernels are generated in sequence.

Change-Id: I55a12a5bc0de3c4eca9ba773a5cf20e2b104a0ee
Signed-off-by: Raghavendra Rao Ananta <rananta@codeaurora.org>