Allow HOSTCC to be set separately from CC
The Linux kernel kbuild will modify CC with additional flags such as
--target, --prefix, --gcc-toolchain, however when building U-Boot with
kbuild, these are expected to be passed as part of CC. However, the
options apply only to CC, not HOSTCC, so we need to be able to control
the setting of these variables independently.

If HOSTCC is not set but CC is, the original behavior is retained.

Bug: 162801433
Change-Id: I4f591183e0f8bb2adc99f90b525198233750ae63
Signed-off-by: Alistair Delva <adelva@google.com>