ABI: Turn on KBUILD_SYMTYPES=1 in build_abi.sh
The genksyms utility is responsible for determining symbol CRCs when
MODVERSIONS is enabled. It uses syntactic type information to
fingerprint each variable and function exported from a module.

The KBUILD_SYMTYPES environment variable causes the -T (--dump-types)
option to be passed to genksyms resulting the creation of a .symtypes
file per source (.c or .s) file. These files can be helpful when
determining the cause of a CRC change.

This commit enables KBUILD_SYMTYPES by default for ABI builds in order
to reduce the time such investigations take. It also reflects this
behaviour in the ABI documentation.

Signed-off-by: Giuliano Procida <gprocida@google.com>
Change-Id: I5d0f0826d70b5e26fa8aa6244e96ff0cec04b07f