build.sh: reset a bunch of variables kbuild would
so that they may be used throughout the script.  The only current
consumer is DO_NOT_STRIP_MODULES.  This could be simplified to only
check for LLVM=1 there, but I assume this pattern of trying to use these
variables in this script may continue. To be safe, just set them all
now. YAGNI? Maybe.

Bug: 169398087
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Change-Id: I36e49c258b7be5ace2fb4cbc79c92be4068392b9