brunch: Set default KERNEL_DIR
In commit e33d6141981f ("_setup_env.sh: set default value for
KERNEL_DIR"), the default value for KERNEL_DIR is set to the location of
the build.config. brunch creates a build.config in root directory, which
is not the kernel directory. In the generated build.config, set
KERNEL_DIR to the folder where modules.list is found.

Change-Id: Ib8daed8d475b72794242f92462fb259456f8aa70
Signed-off-by: Elliot Berman <eberman@codeaurora.org>