build: introduce DISABLE_LTO_CFI_CONFIGS option
LTO is good to shrink the kernel size but it takes long time to build
the kernel image. This patch is providing an option to disable LTO and
CFI options while development.

The options to disable LTO and CFI is derived from [1].

$ DISABLE_LTO_CFI_CONFIGS=1 BUILD_CONFIG=common/build.config.gki.aarch64 build/build.sh

or you can set this in build.config.

DISABLE_LTO_CFI_CONFIGS=1

[1]: https://source.android.com/setup/build/building-kernels#customize-config

Bug: 173434774
Signed-off-by: Chanho Park <chanho61.park@samsung.com>
Change-Id: I77fe1e5a4b0f0b46295cbf1cd4858caf7183a60a