merge_dtbs.py: Add check to ensure DTBOs apply to DTBs
Check that the generated DTBOs apply to the base DTBs. Use msm-id to
match the DTB and DTBOs. This mimics what could be done during Kbuild
time, but is no longer possible since the base DTB is not available
to techpacks. ufdt_apply_overlay behaves similarly to how bootloader
applies DTBO to base DTB.

Change-Id: Ifa8990ad45a623c5d7d016775ad6a15e914ff1e5
Signed-off-by: Elliot Berman <eberman@codeaurora.org>