create_modules_staging: Support toybox grep semantics
When hermetic toolchain is enabled, toybox grep behaves differently than
GNU grep when a pattern file is empty. In GNU grep is provided an empty
pattern file, then it is equivalent to passing no patterns. With toybox
grep, it is equivalent to passing an empty pattern (""). We need the "no
patterns" behavior, so don't pass the blocklist patterns flag when there
is no modules blocklist.

Bug: 178191979
Change-Id: I3c0a25dfb5ab6632146ce7338fe2b4d2419bfd44
Fixes: a1c93623b928 ("create_modules_staging: Correct grep usage in TRIM_UNUSED_MODULES")
Signed-off-by: Elliot Berman <eberman@quicinc.com>