download_from_ci creates repo branches
Currently download_from_ci creates new commit messages in the Android
repository and, in most of the cases, it might happen that these
projects are in detached states and the commit to be unreachable.
This is annoying especially when multiple projects are updated in a
single download_from_ci run.

This patch introduces the "-r" or "--repo-branch" flag that, if enabled,
creates repo branches in the Android projects affected by
download_from_ci changes.
All the branch names have the same format:

  kernel-update-b<BUG>-<BUILD_ID>

and simplify the upload phase, that can be easily performed from the Android
repository with:

  repo upload -t --br kernel-update-b<BUG>-<BUILD_ID>

This command (that is also printed as a suggestion whenever
download_from_ci creates a new commit) uploads all the Android projects
matching the branch name and, with the "-t" flag, groups them within the
same Topic, allowing TreeHugger to run its presubmits integrating all
the changes in that topic.

Test: run script and upload builds
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I8a3643cd4dc71f1331436a4b2ed097c046b874e9