merge_dtbs.py: Kernel -> Techpack superset handling
Current merge_dtbs.py script does not handle the case where the kernel's
devicetree may apply to a superset of a techpack's devicetree.
Specifically, a kernel devicetree may advertise support for Boards A and
B jointly, whereas vendor devicetree may need separate devicetrees for
board A and board B. To handle this, merge_dtbs.py should split the
kernel devicetree into 2 (or more) "final" devicetrees which would have
different techpacks applied according to the techpack needs.

When splitting devicetrees, we need a unique filenames for the resulting
devicetrees. For now, let's combine the filenames of the merged
devicetrees.

The current script didn't have scalable way to support this, so rather
than hack it in, let's switch to a more object-oriented way of tracking
devicetrees and how they are merged with the techpacks.

Change-Id: I855820d8c2174e423eb3646a08bbabc5dc576ded
Signed-off-by: Elliot Berman <eberman@codeaurora.org>