build.sh: Move config save after mrproper
When running with a few more integration tests, I discovered that
include/config/auto.conf was not in sync with a restore .config.before.

When running with SKIP_MRPROPER unset, the config.before was restored as
reasonably expected. However, Kbuild uses auto.conf for configuration
values at build-time. For concrete example, this command is used:
  LTO=thin BUILD_CONFIG=common/build.config.gki.aarch64 ./build/build.sh
If the build was already run, .config would have thin LTO in the
.config. In the incremental build, auto.conf is initialized with full
LTO as specified in gki_defconfig. Later, we restore .config.before
since it matches the original .config. There is now nothing to have
Kbulid realize that auto.conf is out-of-sync with the .config since the
timestamps are aligned.

To avoid this scenario, move the .config checkpoint after mrproper
step. Now, the .config will be cleaned and compilation proceeds as
expected.

Bug: 179648610
Fixes: c84e95adb48a ("build.sh: Avoid unnecessary touching of the kernel
config by build.sh")
Signed-off-by: Elliot Berman <eberman@quicinc.com>
Change-Id: Iba5aa1b1c5e558d05bbbd49fc333be1e3c18cb5c