add synchronize_owners utility
This helper synchronizes OWNERS directives between a main branch and a
tracking branch. In the directories in which the main branch has an
OWNERS file, the tracking branch's equivalent OWNERS file will have a
directive to include the main branch file. Existing OWNERS file contents
are preserved by prepending the include directive if absent.

Example usage (executed in an common-android-multi repo checkout):
 $  build/synchronize_owners common-mainline/ android-mainline common12-5.10/

Bug: 184248201
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I41b0e0e8ff20b45ef1f1e8641e1107af95bc7c85