build_abi.sh: Disable mixed build
Explicitly disable mixed build when running build_abi.sh. The ABI
comparison when doing a mixed build should be done against the device
kernel, not the GKI/reference kernel. The GKI kernel is usually the ABI
which the device kernel needs to be compatible with, so doing a
comparison with a vmlinux generated from that tree is not particularly
interesting.

Bug: 178469391
Change-Id: I91ecd07ce9fc4000d4ffc2388c008552e826be7b
Signed-off-by: Elliot Berman <eberman@codeaurora.org>