build_module.sh: Introduce KERNEL_KIT
The current version of build_module.sh has a limitation that it requires
a complete pre-compiled kernel source tree. This is not ideal as it
requires a complete kernel build available inside Android builds.

Introduce support for a new variable, KERNEL_KIT, which allows the
caller to specify a prebuilt folder which contains the necessary
pre-compiled artifacts (.config, Module.symvers, and host tools). Then,
"make modules_prepare" is run to create necessary generated header files
and Kbuild host tools (like fixdep).

With this change, Android can compile its out-of-tree kernel modules
like this:

  cd kernel_platform
  OUT_DIR=${ANDROID_PRODUCT_OUT}/obj/DLKM_OBJ \
  KERNEL_KIT=../device/qcom/taro-kernel \
  EXT_MODULES=../vendor/qcom/opensource/audio-kernel \
  ./build/build_module.sh

This change is also backwards compatible with a userspace that doesn't
know how to use KERNEL_KIT. In that case, the end-to-end behavior is the
same as the original build_module.sh.

Change-Id: Ib23444572c4aa1b647a1213d49375cce998fc83a
Signed-off-by: Elliot Berman <eberman@codeaurora.org>