android: prepare_vendor.sh: Support kernel kit build_module.sh
Specify KERNEL_KIT and OUT_DIR to align with build_module v2. With
build_module v2, out-of-tree modules can be compiled "directly" in the
desired output location: ${ANDROID_PRODUCT_OUT}/obj/DLKM_OBJ. This is
achieved by creating the real kernel platform output folder (for
out-of-tree modules) inside
${ANDROID_PRODUCT_OUT}/obj/DLKM_OBJ/kernel_platform. As far as
Android-built out of tree modules are concerned, kernel platform was
compiled into that folder.

This change has the added benefit of being able to remove the previous
hack of creating a symlink of "la" to the root of the Android build
environment. That comes at a cost of no longer being able to use an
kernel platform source tree that lived outside of Android's workspace.
However, that wasn't a well supported use case anyone and there are no
known users of it.

Change-Id: I923fb4d9ece28e998c463b8550cb7c39231e26c5
Signed-off-by: Elliot Berman <eberman@codeaurora.org>