Kleaf: Retire direct use of build/build.sh
Instead, directly invoke make targets for the kernel build. Later, more
steps can be introduced to build header packages, ABI representations
and modules, etc.

At this point, we have all prerequisites to build kernels without
build.sh. We still lack a lot of features of it, but we do not require
it any longer for the targets we have.

Bug: 192656402
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: I683cd0874d48f0859838af34357648bdc6b502a7