Kleaf: kernel_env: use providers to propagate how to use the env
Let kernel_env return a provider "KernelEnvInfo" that defines the
dependencies and the steps to execute for downstream rules. That way we
can make changes to the environment at the definition and they get
propagated to the users as typed information.
In particular, collect all tools needed to use this environment as
labels while allowing a custom toolchain version to be specified (that
also propagates). Further, define a setup script that is supposed to
prefix any downstream commands to setup the environment.

For now, this creates a bit of duplication that we can clean up later
once all downstream rules make use of this provider.

Bug: 192656402
Signed-off-by: Matthias Maennich <maennich@google.com>
Change-Id: Iebcabc246fff7636417af392b2f3c15b89c8af29