Kleaf: kernel_module output to package output directory.
Previously, files of module foo is written to
  <package_output_dir>/foo

After this change, write to
  <package_output_dir>

directly instead, so that its behavior is the same as
other Bazel rules (e.g. gendir).

This allows kernel_module.outs to be an attr.output_list.

(Note: for kernel_build targets this still builds to
the subdirectory foo.)

Test: builds a module
Bug: 193896689

Change-Id: Ib6efb859af661c1b3090aaf90d322f601962c7de