build.sh: Introduce BOOT_IMAGE_FILENAME
Replace the hard-coded file name "boot.img" by a flag allowing
BUILD_CONFIG to overwrite the name of the output of mkbootimg; as the
default value of BOOT_IMAGE_FILENAME is the string it replaces, this
change should not affect existing uses of build.sh. This is necessary
for Android Partitions using the Boot Header format other than boot.img
itself (e.g. pvmfw.img).

Bug: 194265608
Change-Id: I69b5b29326ef0c7df7348bcb31e34b7f754d5466