Kleaf: add remote build capabilities
This introduces --config=remote support for kleaf builds. Most kernel
build actions are not yet compatible with remote builds, but this commit
sets up the configuration to work on that.

Remote building requires an RBE one-time setup. With authenticated
access to a remote build instance, usage boils down to having valid
values for project and instance defined in ~/.bazelrc or other locations
bazel looks for configuration:

 | build --project_id=<project>
 | build --remote_instance_name=<rbe_instance>

Change-Id: I2584230354f2a34d8095be406bd4cd22c35bafd7