kleaf: //build:host-tools -> //build/kernel:host-tools
This is part of the build/ -> build/kernel migration.

The //build/kernel:host-tools module already exists,
so we migrate all usages to it.

Remove the //build:host-tools module to prevent further
usage with a conditional. Right now, build/BUILD.bazel
is still a symlink to build/kernel/BUILD.bazel. Once
the symlink is removed, the conditional can be deleted.

Bug: 217630659
Test: TH

Change-Id: I2e7260c02d3a1d53833584a863ea5051efb2a698