kleaf: Delete KernelFilesInfo.
The provider was created as an alternative to
DefaultInfo that requires a user to use kernel_filegroup
instead of filegroup.

However, such obstacles are unncessary. With the provider,
the _kernel_build_rule had too many interfaces it implements:

- DefaultInfo
- _KernelBuildInfo
- _KernelEnvInfo
- OutputGroupInfo
- KernelFilesInfo
- _kernel_build_aspect

This makes it hard to replace kernel_build targets with kernel_filegroup
targets, because kernel_filegroup needs to implement all
these interfaces as well.

Delete the KernelFilesInfo interface and merge it into
_KernelBuildInfo. This allows us to replace kernel_build
with filegroup in the future.

Test: TH
Bug: 218761417
Change-Id: I135162603a7f06158a7f440f93599efc1679a084