kleaf: Split build_boot_images to build_boot and build_vendor_boot.
The two booleans are clearer than the one boolean
build_boot_images. In addition, if build_boot is True
but build_vendor_boot is False, the build command
adds SKIP_VENDOR_BOOT=1.

This also removes the dependency to SKIP_VENDOR_BOOT=1 in build.config,
which is one step away from having build.config's that are only used in
build.sh.

Test: TH
Test: manually turn off build_vendor_boot and build, see that
  vendor_boot.img is not built.
Bug: 222587028

Change-Id: Iccee088b5b0e0b908df8ea2a268a1021c5250a8f