ANDROID: move dump_kernel_abi out of AbiTool classes
Moving this method out essentially bakes in the (current and only)
choice of calling `abidw` and `abitidy`.

There is likely no chance of a drop-in compatible replacement for
`abidw` in the medium-term at least and, in any case, decisions about
which tool to run should probably not be exposed to external users.

Bug: 226567130
Change-Id: I130ce0b081ed90d8eb0e51e4b2a5a75e535011ae
Signed-off-by: Giuliano Procida <gprocida@google.com>