kleaf: search_and_cp_output.py create dirs for output names with slash.
For outputs that are not declared directly in Bazel, e.g.
when copying outputs to a declared directory, also
create the directory structure in the destination directory.
Because the relative path is not declared in Bazel, Bazel won't
create the directory for us.

In particular, when copying unstripped modules from ${OUT_DIR}
to unstripped/, if `module_outs` states a relative path like
`net/wireless/cfg80211.ko`,
this creates `unstripped/net/wireless` before copying to
 to `unstripped/net/wireless/cfg80211.ko`

This is a different behavior than build.sh, where the
UNSTRIPPED_MODULES list must be a list of basenames, and the
unstripped/ directory is flattened.

Test: TH
Bug: 227176212

Change-Id: Iaa0e7375fe3bbcc151076a00c7db0d2a35fe6d27