kleaf: kernel_module executes even with no outs.
Add a "check_no_remaining" file to the default
outputs of a kernel_module target so that, even when
the "outs" list is empty, the KernelModule action
is executed, and the check_declared_output_list
script is executed to make sure the outs list is
up to date.

Because the default outputs of kernel_module is
changed, modify rules that depend on it to use
a different list instead.

Test: bazel test [module]_test
Bug: 230799504
Change-Id: Id0a7da9ad6b3fb38fa3a0cc34d4a2b1dce7e57b0