kleaf: kernel_module define siblings
Define sibling versions of external kernel_module
that is built on top of the corresponding sibling version
of the kernel_build.

These two targets (_notrim and _with_vmlinux) are used
in ABI monitoring. ABI targets that uses the notrim
or with_vmlinux variant of the kernel_build will use
the notrim or with_vmlinux variant of the external
kernel modules as well, respectively, in the following
CLs.

Test: TH
Bug: 231480193

Change-Id: Ia02903023e9a3138d7bf0bc2a8be0ee845eb7188
Signed-off-by: Yifan Hong <elsk@google.com>