download_from_ci: get_binary_kernel_version works for arm64 and x86
So far, get_binary_kernel_version worked only for arm64 and not for
x86_64. This was because the function was implemented by grep'ing a
specific string literal 'Linux version ' embedded somewhere in the file.
Unfortunately, that string literal doesn't exist in x86_64.

This change fixes the problem by following the x86 kernel image
format https://www.kernel.org/doc/Documentation/x86/boot.txt to locate
and extract the version string correctly.

The code then falls back to grep'ing the 'Linux version ' for the arm64
kernel.

Bug: N/A
Test: run download_from_ci
Change-Id: I05f19a78377a6a07ba96ee894a0796af461f56df