kleaf: Remove toolchain_version from default outputs of //common:kernel_aarch64_downloaded
The file was added to DefaultInfo of kernel_build
only to allow downloading toolchain_version from
build servers so that kernel_filegroup works.
kernel_filegroup also contains this file in
DefaultInfo to be consistent with kernel_build.

However, this caused a conflict when the device's kernel_build
and the downloaded GKI prebuilts (//common:kernel_aarch64_downloaded)
are both added to a copy_to_dist_dir rule with flat=True (even though
the two files should have the same content).

Hence, drop it from the DefaultInfo.

Test: TH
Test: manually build raviole with --use_prebuilt_gki
Bug: 244619402

Change-Id: I9487da0989bccb2604448f34e0d2da8d849bd596