kleaf: Allow setting the Makefile for a kernel_module()
This allows one to define a kernel_module() starlark outside of the
actual kernel module's directory by setting kernel_module.makefile to
the location of the external module's Makefile. This is equivalent to:

  make M=<path-to-module-sourc> ...

Currently, we have to hard code the outputs of a given kernel_module(),
which restricts the usage of that kernel_module() to a specific
defconfig configuration. To allow for re-use of an external module
that's output varies based on the defconfig, we can define the
kernel_module() starlark in a separate project per defconfig
combination. This also helps with separating device-specific
configurations from the driver source code.

Bug: 257343887
Change-Id: Id754f38f5b23fdd5a1d8e8603e637ed52cfda284