android: prepare_vendor: Setting default KERNEL_VARIANT
When no specific kernel variant is passed on the command line, the
current logic depends on the build.sh script to derive a default
variant from the target's build.config.

The top-level Bazel build script does not have such behavior and
needs the variant to be passed in.

Add logic to derive the default kernel variant from prepare_vendor
so that it can be passed into build_with_bazel.py.

Change-Id: I2af2c631840315506f44d13aa17366c78cd70c49
Signed-off-by: John Moon <quic_johmoo@quicinc.com>