prepare_vendor: Add check for unsigned_abl.elf before use
Currently, the script assumes the existence of abl-userdebug/
unsigned_abl.elf when running a diff check. When building with Bazel,
this file does not exist, so an error message is printed:

diff: abl-userdebug/unsigned_abl.elf: No such file or directory

Update the script to check for the existence of the file prior to use.
The flow control logic will operate the same way, just the error
message will be suppressed.

Change-Id: I6c1052c6e28c1a0e1497ad2effcf81e70aaad184
Signed-off-by: John Moon <quic_johmoo@quicinc.com>