kleaf: Improve hack for fixdep.
In the previous code, after we drop the hack that also
restores fixdep, android14-5.15 local builds break.

For sandbox builds, `modules_prepare` and other downstream targets
has the following lines in auto.conf.cmd:

  ifneq "$(srctree)" "<sandbox execroot>/common"
  include/config/auto.conf: FORCE
  endif

The sandbox execroot is different for kernel_config and for
modules_prepare, causing auto.conf.cmd to be rebuilt:

  SYNC    include/config/auto.conf.cmd

... which in-turn triggers rebuild for fixdep before building
other hostprogs (gen-hyprel in this case).

However, for local builds, auto.conf.cmd has the following lines:

  ifneq "$(srctree)" "<repository root>/common"
  include/config/auto.conf: FORCE
  endif

The value of srctree does not change, hence auto.conf.cmd is
not rebuilt, which does NOT cause fixdep to be rebuilt. Then,
other hostprogs (gen-hyprel in this case) fails because rules
using if_change_dep does not depend on scripts/basic/fixdep.

This change replaces the hack that stores fixdep binary with
a better hack. It forces all downstream targets to ALWAYS re-build
include/config/auto.conf, which will cause fixdep etc. to be rebuilt.

Alternatively, this could be fixed by fixing EVERY make target
that uses if_change_dep in the kernel source tree to depend on
scripts/basic/fixdep properly. However, this seems to be a much harder
task to do.

Bug: 263415662
Bug: 267385639
Test: integration test

Change-Id: Id510d90a11dee81b6637505f599211a4ac0a961f