prepare_vendor.sh: Add Bazel symlink assurances
Currently, if the Bazel build extensions do not exist or are broken
links, build_with_bazel.py will recreate the links and point to the
default extensions.

In prepare_vendor.sh, in a case where ABL is recompiled, but the
kernel is not, the extensions may be broken (which causes a build
failure).

Make sure the links are set to the defaults in case they're broken
or non-existent.

Change-Id: I86ed08efbf1f2dd44009bbe6b40f99f590f98566
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>