kleaf: Add symbol violations buildtime check
Add a call to abi/check_buildtime_symbol_protection to ensure
undefined symbols in unsigned modules are accounted for as
either listed in the symbol list (GKI public interface) Or
exported by another unsigned module (vendor interface).

This is the re-land of the original aosp/2410926 with original
commit message which was reverted due to broken targets caught
by this check.

Bug: 269488249
Change-Id: Ie4a95c565f84e54a931505147fa9fe6e485a645d
Test: TH
Signed-off-by: Ramji Jiyani <ramjiyani@google.com>