Reland "kleaf: Add back force_disable_trim bool_setting."
This reverts commit 1c1edcada4ba76575ed5b72807b6e34c46ab6d04.

Instead of relying reading trim configuration from
a string_setting trim_nonlisted_kmi_setting,
read from ctx.attr.trim_nonlisted_kmi directly.
The attribute is always select()'ed based on
force_disable_trim, so that it is affected by
notrim_transition.

With this change, because the scope of ctx.attr.trim_nonlisted_kmi
only affects the current target, for regular builds (that
doesn't have notrim_transition enabled), the base_kernel is not
affected by the trim_nonlisted_kmi setting of the device kernel.

Implementation details: _trim_nonlisted_kmi_transition_impl is cleared
because of the following error:

  Error: No attribute 'trim_nonlisted_kmi'. Either this attribute
  does not exist for this rule or is set by a select. Starlark rule
  transitions currently cannot read attributes behind selects.

The setting is already unused and is pending removal,
so just set the function to no-op for now to silence this error.

Reason of revert: relanding change
Bug: 270670362
Bug: 269674780

Change-Id: I6a1f7a9ca26c034b19c41453fc55eedf74628407