kleaf: Add KernelEnvInfo.run_env.
This is like the regular KernelEnvInfo except that
it is for being used in the execution phase for
bazel run X_config.

The environment is not suitable for kernel build; it
does not set up kconfig_ext, dtstree, source_date_epoch,
etc. This is just a bare minimum environment for running
the Kleaf equivalent of config.sh.

Test: TH
Bug: 270996321

Change-Id: I60b78160b8f217063bd0688e7fe086ba79f72ae5