kleaf: Fix setup script for "run X_config". (second attempt)
This reverts commit 18e439378c1e154c7c061216532e31005bf344e4.
In addition, this change incorporates a few additional fixes.

- Fixed comment on test_config_uapi_header_test
- Fix logic that finds DEFCONFIG so that x86_64 works

riscv is excluded because NDK_TRIPLE is not set, causing
--sysroot to be unset, causing CONFIG_UAPI_HEADER_TEST to go away.
This is tracked in b/271919464.

Here's the original commit message.

Before running e.g. //common:kernel_aarch64_config -- menuconfig,
the script should use KernelEnvInfo.run_env instead, because
that's the proper environment for execution phase.
The run_env environment also ensures --sysroot in USERCFLAGS
are restored properly.

Also, the script does not enter BUILD_WORKSPACE_DIRECTORY at the beginning.
Instead, it is executed in execroot like any other runnable executables. The
script may run `savedefconfig` as part of POST_DEFCONFIG_CMDS,
which overwrites the symlink in the execroot (e.g.
common/arch/arm64/configs/gki_defconfig). The script ensures
that the change is propagated to the real source file, and reinstate
the symlink so a future `bazel run` can be executed properly.

This change is necessary so that CONFIG_UAPI_HEADER_TEST is not
deleted from gki_defconfig.

This CL also adds integration tests to ensure the fix does not regres.

Bug: 270996321
Bug: 271919464
Test: bazel run //common:kernel_aarch64_config -- olddefconfig
  does not delete CONFIG_UAPI_HEADER_TEST
Test: TH (integration test)

Change-Id: I60452c818e8f62abaaf04ac946a2c431470aa64f