kleaf: Add ddk_config.
This is a way for developers to specify module-specific Kconfig
and defconfig.

Both kconfig and defconfig are optional.

- If both are missing, the .config etc. from kernel_build is used.
- Sets KCONFIG_EXT_PREFIX properly to refer to a generated Kconfig.ext file.
  The generated Kconfig.ext file always source-s the Kconfig.ext
  from kernel build.
- If ddk_config.kconfig is set, the generated Kconfig.ext file also
  source-s this file.
- If defconfig is provided, it is merged into .config.
- If any of kconfig / defconfig is provided, the $OUT_DIR/include
  is regenerated based on the (possibly new) .config file and
  with KCONFIG_EXT_PREFIX set.

Note that in this change, instead of a single kconfig / defconfig
file of this ddk_module, a depset is used in the scripts. This is
to prepare for follow up changes that includes Kconfig / defconfig
of depending modules. See follow up CL.

Bug: 264293217
Test: bazel test //common-modules/virtual-device/kleaf_test
Change-Id: Ie99f15abc14dac8183a83ffea65b3814c3c0ac92