qcom_symbol_list_update: Add -a flag for manual additions
Currently, the script will only add symbols that are used by the
msm-kernel build. Sometimes, we may want to add a symbol manually
to speed up processes or avoid a headache of moving code around just
to enable use of certain symbols.

To facilitate manual additions, add an -a flag where the user can add
arbitrary symbols to the list and have them properly sorted/deduped.

Change-Id: If441ea9fe74213c923641f5711c413730e187ab4
Signed-off-by: John Moon <quic_johmoo@quicinc.com>