kleaf: raw_kmi_symbol_list / combined_abi_symbollist allow label with no file.
allow_single_file requires that, if a label is provided
to the attribute, it must contain exactly one file. It
cannot contain 0 files. This change makes the cases
more generic by allowing labels with 0 files when no
abi symbol list is provided to kernel_build, so
kernel_build.kmi_symbol_list may be made configurable
in a follow up CL.

Test: TH
Bug: 283159637
Change-Id: Idee4679cd1d33f3ebcc2a31b1c7148a9a925bfa8