kleaf: Add flag to ignore missing repo projects
Currently, the workspace_status_stamp.py script will perform various
operations on all projects defined in the repo manifest.

In some cases, the project directories may not be present. When this
happens, the build currently fails. Users should have the option to
change this behavior.

Add a flag, --ignore_missing_projects, which will ignore conditions
where a project is listed in the repo manifest, but does not exist in
the workspace.

Bug: 284943448
Change-Id: I5c58952aa48c4637f733d3e2bbb0d15f56606626
Signed-off-by: John Moon <quic_johmoo@quicinc.com>