build_module: Add support for more deep nesting of packages
Currently, DDK build will only be attempted if the Bazel package's
BUILD.bazel file is found in the module directory or one level above.

Some packages are more deeply nested in the package directory
structure, so this method does not work.

Instead, search up from the module directory until a BUILD.bazel file
is found. If we hit the top-level workspace, we fall back to the
legacy build.

Change-Id: If6c3883575066c2da144a6398cefdb0beba9e157
Signed-off-by: John Moon <quic_johmoo@quicinc.com>