kleaf: Add --no-group to all rsync calls
Currently, some rsync operations result in the following error
message:

rsync: chgrp "<sandbox>/<file>" failed: Invalid argument (22)

With some syscall analysis, this issue appears to be driven by the
Linux namespacing used by Bazel. Inside the sandbox, the group rsync
is trying to assign to the file does not exist, so the syscall fails.

To avoid this issue, pass "--no-group" to rsync. That way, it will
not attempt to set the group. This will work correctly by default
anyway as the build will create the file with the user's default
GID.

Change-Id: Iea49985ff465568b49d2bf082868cfa0ea95351b
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Linux Image Build Automation <quic_ibautomat@quicinc.com>