build_module: Switch to using build_opts.txt to discover build flags
Currently, we parse the output kernel .config file to determine which
LTO flag to pass to the DDK build. As more flags are added to the
kernel build, we need to find a better way to make sure the flags
passed to the kernel and the flags passed to the DDK build are the
same.

We've added an artifact to the kernel build, build_opts.txt, which
contains the Bazel command line flags passed to the kernel build.
Using that file to supply the flags to the DDK build , we can
guarantee the flags stay in sync between the two.

Change-Id: I647151ab11937263e222d62f7bacba7f1143e4a6
Signed-off-by: John Moon <quic_johmoo@quicinc.com>