qcom_symbol_list_update: Add local modification option
Currently, the qcom_symbol_list_update.sh script is able to create a
change which updates the upstream ACK ABI list. In some cases, we may
want to update the msm-kernel ABI list directly instead of syncing
the ACK list first.

Add an option (-L) to the script which modifies the msm-kernel list
directly to allow for downstream updates.

Change-Id: If8b7bf68db5e078047c3fb46d6cdef790add3a18
Signed-off-by: John Moon <quic_johmoo@quicinc.com>