build_module.sh: Remove xargs -a argument
Some versions of xargs do not support the -a option which is used to
pull arguments directly from a file. When the build is run with these
versions, the following error is seen:

xargs: Unknown option 'a' (see "xargs --help")

To support these versions of xargs, change "xargs -a <file>" to
"cat <file> | xargs" which has the same effect.

Change-Id: Id9d3aada8341f97aa64ef259ac9c51d8bdc3e4c3
Signed-off-by: John Moon <quic_johmoo@quicinc.com>