kernel_sbom: add option to pass version file
Add an option to alternatively pass a file name via --version_file that
contains the version string instead of passing the version as a string.
The kernel already has a `kernel.release` file for that purpose. By
letting the program open it directly, we save a subprocess call to cat,
but moreover, we do not need to depend on the hermetic toolchain just
for cat.

Bug: 279663374
Change-Id: I312e4246e736ca647c465336ca2d45c74f7de31a
Signed-off-by: Matthias Maennich <maennich@google.com>