kleaf: Add ddk_uapi_headers rule
It's common for external kernel modules to publish UAPI headers which
need to be sanitized prior to delivery to userspace. Currently, the
DDK has no way to package UAPI headers this way.

Add a new rule, ddk_uapi_headers(), which allows external modules
to package their UAPI headers for userspace. For example:

    ddk_uapi_headers(
       name = "my_headers",
       srcs = glob(["include/uapi/**/*.h"]),
       out = "my_headers.tar.gz",
       kernel_build = "//common:kernel_aarch64",
    )

Bug: 313898637
Bug: 326633656
Change-Id: I1b7b59def1bf0b99b1073600e1324c21f82996e8
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>
[isaacmanjarres: Used config_env_and_outputs_info and
get_setup_script, since KernelSerializedEnvInfo is not
available on this branch. Also added module_scripts
and part of the inputs for running the rule.]