kleaf: Fix KBUILD_MIXED_TREE for --incompatible_sandbox_hermetic_tmp.
Previously without the flag, the TreeArtifact contains symlinks to
files under execroot, which is allowed.

With the flag, the symlink points to a file in the sandbox, which
is no longer valid (the sandbox is teared down) when the TreeArtifact
is validated.

Change it to a cp so that it is straightforward to humans.

Performance: There are no obvious performance regression associated with
this change. The KbuildMixedTree action is on the critical path.
This change reduces the time for the action from 4.1s to 3.6s in a
one-time experiment.

Test: TH
Bug: 319893821
Bug: 319077326
Change-Id: I63f2364e0ffaa57a7ae57a1f7730f4b510ae3a8e
Signed-off-by: Yifan Hong <elsk@google.com>
(cherry picked from commit b626eb43be646d8efdfb57e7d832da39fd81284a)